Added a GameStateManager and code to change spawn points
GameStateManager will maintain things between scenes like the candles
that have already been lit. The barrels are not maintained but I think
it’s ok for they to pop up again (but we can change that if needed)

Also figured out how to change spawn point depending on which room
player is coming from.

For every door, the WallWithDoorPrefab has a SpawnerPoint empty object
that is where the player will spawn at (you can change its position if
there is something in from of the door like the barrels on the first
room “GameScene”).

On the ExitDoor script attached to it there is a field called “Spawn
Name” that you have to set according to which room player is going to
(so for the first room there is only one door, and that field is called
“1and2”). Right now it connects to one door in the hallway which has
the same name, so they are connected.

If this doesn’t make sense just ask me :P